+++ tests/terranix-tests.nix
@@ -81,4 +81,9 @@
     file = ./terranix-tests/14.nix;
     outputFile = ./terranix-tests/14.nix.output;
   }
+  {
+    text = "meta: allow _meta passthru without affecting json";
+    file = ./terranix-tests/15-meta.nix;
+    outputFile = ./terranix-tests/15-meta.nix.output;
+  }
 ]